Move transport specs into integration directory
These tests all rely on having a target available to run them against,
which makes them much more like integration tests than unit tests. With
this change, only one set of tests in the `spec/bolt` directory still
relies on an external service (some of the PuppetDB client tests).
