+++ spec/integration/transport/winrm_spec.rb
